补充：Sort类中 堆排序 Heap() 的设计要点：   1、基本思想：将待排序的序列以BuildHeap()构成一个大/小顶堆。此时，整个序列的最大/小值就是堆顶的根结点。并将它与堆数组的堆末尾元素交换，在堆末尾形成有序区(最大/小值)。然后将剩余的n-1个序列重新构造成一个堆，这样就会得到n个元素中的次小/大值。如此反复执行，便能得到一个有序序列；   2、一个创建最大/小堆的过程分析：首先按数组的数值顺序创建一个完成二叉树，再从不是叶子结点的倒数第一个结点比较父结点和孩子结点的大小，按升/降序进行数值交换，再判断这个交换后有没有对孩子结点作为父结点的子树大小排序产生影响，若有对这个子树进一步做堆调整；再对倒数第二个非叶子结点的子树进行上述排序的过程，直至最大值到达根结点；再将根结点与末尾结点进行数值交换；   3、堆排序的时间复杂度为O(nlogn),算法不稳定；   4、代码细节点：      没任何排序操作之前，这个二叉树结点数值的顺序从上至下从左至右的数值排序，与待排序的数组数值顺序一样，根结点编号为0对应的就是数组下标为0的元素；      left() 和 right() 返回的是编号为i值父节点的左右孩子结点的编号，注意不是数值；      BuildHeap() 中：i=(len-2)/2 开始，跳过所有叶子结点，从倒数第一个非叶子结点的子树开始进行堆调整，这个就以main函数中排序的数组对应画个二叉树图就能理解了；      HeapAdjust() 中：lf < heapSize 阻止对数组越界的情况发生，若下标号越界则就不用比较大小了，后面 largest == i，这种情况就是叶子结点的情况，这个就是递归的出口；      Heap() 接口函数中： for将排序出来的最大/最小值与数组的下标为0的元素进行交换，这样循环完毕，数组也就变成有序的了；   5、其它变化：重载 操作符new 中的参数类型 unsigned int 改变为 size_t 类型，因为QtCreator版本改为5.9，不用这个报错；

+++ Object.cpp
@@ -7,7 +7,7 @@ using namespace std;
 namespace WoodLib
 {
 
-void* Object::operator new(unsigned int size) throw()
+void* Object::operator new(size_t size) throw()
 {
     // 编译器会调用这里分配堆内存空间后自动调用相应类的构造函数
     return malloc(size);
@@ -18,7 +18,7 @@ void Object::operator delete(void* p)
     free(p);
 }
 
-void* Object::operator new[](unsigned int size) throw()
+void* Object::operator new[](size_t size) throw()
 {
     // 当用new Test[5]时，只须传入数组元素的个数
     // 编译器会向operator new[](...)函数的参数
