+++ Object.h
@@ -2,6 +2,8 @@
 #ifndef _OBJECT_H_
 #define _OBJECT_H_
 
+#include <cstdlib>
+
 namespace WoodLib
 {
 
@@ -10,9 +12,9 @@ class Object
 public:
     // 以下四个重载函数用于统一不同编译器new失败时的结果不同的问题
     // throw()表示不抛出异常，即如果申请内请失败时，统一返回NULL而不抛异常
-    void* operator new(unsigned int size) throw();
+    void* operator new(size_t size) throw();
     void operator delete(void* p);
-    void* operator new[](unsigned int size) throw();
+    void* operator new[](size_t size) throw();
     void operator delete[](void* p);
 
     // 考虑其子类要比较两个对象是否相等，重载之(重载==运算符)！
